--- a/.travis.yml
+++ b/.travis.yml
@@ -1,5 +1,8 @@
 language: python
 python: "3.5.5"
+
+cache: pip
+
 services:
   - postgresql
 addons: