--- a/tests/factories/partnership.py
+++ b/tests/factories/partnership.py
@@ -1,8 +1,7 @@
 import factory
 
 from partnership.models import Partnership, PartnershipTag, PartnershipType
-from .contact import ContactFactory
-from .partner import PartnerEntityFactory, PartnerFactory
+from .partner import PartnerFactory
 from .partnership_year import PartnershipYearFactory
 
 __all__ = [
@@ -23,34 +22,23 @@ class PartnershipFactory(factory.DjangoModelFactory):
         model = Partnership
 
     partner = factory.SubFactory(PartnerFactory)
-    partner_entity = factory.SubFactory(
-        PartnerEntityFactory,
-        partner=factory.SelfAttribute('..partner'),
-    )
     partnership_type = PartnershipType.MOBILITY.name
 
     ucl_entity = factory.SubFactory(
         'base.tests.factories.entity.EntityFactory',
+        organization=None,
         country=factory.SelfAttribute('..partner.contact_address.country'),
     )
 
-    author = factory.SubFactory('base.tests.factories.person.PersonFactory')
-
     @factory.post_generation
     def tags(obj, create, extracted, **kwargs):
-        if create:
-            if extracted:
-                obj.tags.set(extracted)
-            else:
-                obj.tags.set([PartnershipTagFactory(), PartnershipTagFactory()])
+        if create and extracted is not None:
+            obj.tags.set(extracted)
 
     @factory.post_generation
     def contacts(obj, create, extracted, **kwargs):
-        if create:
-            if extracted:
-                obj.contacts.set(extracted)
-            else:
-                obj.contacts.set([ContactFactory(), ContactFactory()])
+        if create and extracted is not None:
+            obj.contacts.set(extracted)
 
     @factory.post_generation
     def years(obj, create, extracted, **kwargs):