… tests/views/partnership/test_partnerships_list_view.py