On a 1300+ objets créés par les Factory rien que pour le test de la liste des partenariats (test le plus lent avec ~10s de `setUpTestData()`, je me suis dit que ça commençait à faire un peu trop.

 - J'ai supprimé les SubFactory pas trop utile (notamment les ForeignKey à `null=True`)
 - J'ai rationalisé la création d'objets dans `PartnershipsListViewTest`
 - J'ai modifié les factory qui créaient des objets automatiquement en post-generation (pour préférer le manuel)

Bref, en local :

|           | Avant | Après |
|-----------|-------|-------|
| Parallèle | ~66s  | ~37s  |
| Normal    | ~100s | ~73s  |

J'ai également ajouté le cache de pip sur travis, donc les builds devraient être encore plus rapides.